Move to subpackage and move visitor to separate file
